Implement middle-man #delete funtion to avoid bug
Instead of allowing a direct call of the recursive function #delete_node
(previously called #delete), my implementation now hides this function
and exposes a secondary function #delete. #delete calls #delete_node,
but it also assigns the return of #delete_node to the Tree's @root.

This fixes a bug that could occur when trying to delete the root node of
a heavily imbalanced Tree - one where the root node only has one child.
In the previous implementation (and in other implementations I have
seen), the recursive function would succesfully return a new root node,
but would not assign this node to @root. Therefore the Tree object would
remain unchanged.

In my implementation, @root is always re-assigned, thus changing the
Tree object.

This has the secondary advantage of exposing a public #delete method
with only one accepted argument. This helps hides the existance of the
Node class from the public interface of Tree, thus removing a potential
area of user error. A similar tactic could be used on the #insert method
in future iterations.